requirements.txt
appdirs==1.4.3
falcon==1.1.0
gevent==1.1.2
greenlet==0.4.12
gunicorn==19.7.1
## !! Could not determine repository location
iris-admin==0.0.1
packaging==16.8
PyMySQL==0.7.2
pyparsing==2.2.0
python-mimeparse==1.6.0
PyYAML==3.11
six==1.10.0
SQLAlchemy==1.0.11
ujson==1.35
